Extract the queue_job from connector
