+++ connector/__init__.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
 from . import setting
 from . import backend_model
-from . import queue
 from . import connector
 from . import producer
 from . import checkpoint
-from . import controllers
-from . import jobrunner
+from . import related_action
