+++ connector/connector_menu.xml
@@ -8,25 +8,6 @@
         name="Connectors"
         parent="menu_connector_root" />
 
-    <menuitem id="menu_queue"
-        name="Queue"
-        parent="menu_connector_root"/>
-
-    <menuitem id="menu_queue_job_channel"
-        action="action_queue_job_channel"
-        sequence="12"
-        parent="menu_queue"/>
-
-    <menuitem id="menu_queue_job_function"
-        action="action_queue_job_function"
-        sequence="14"
-        parent="menu_queue"/>
-
-    <menuitem id="menu_queue_job"
-        action="action_queue_job"
-        sequence="18"
-        parent="menu_queue"/>
-
     <menuitem id="menu_checkpoint"
         parent="menu_connector"
         name="Checkpoint"
