+++ connector/exception.py
@@ -19,8 +19,16 @@
 #
 ##############################################################################
 
+from odoo.addons.queue_job.exception import (
+    RetryableJobError,
+    JobError,
+)
 
-class ConnectorException(RuntimeError):
+
+# Connector related errors
+
+
+class ConnectorException(Exception):
     """ Base Exception for the connectors """
 
 
@@ -32,50 +40,15 @@ class InvalidDataError(ConnectorException):
     """ Data Invalid """
 
 
+# Job related errors
+
 class MappingError(ConnectorException):
     """ An error occurred during a mapping transformation. """
-
-
-class JobError(ConnectorException):
-    """ A job had an error """
-
-
-class NoSuchJobError(JobError):
-    """ The job does not exist. """
-
-
-class NotReadableJobError(JobError):
-    """ The job cannot be read from the storage. """
-
-
-class FailedJobError(JobError):
-    """ A job had an error having to be resolved. """
-
-
-class RetryableJobError(JobError):
-    """ A job had an error but can be retried.
-
-    The job will be retried after the given number of seconds.
-    If seconds is empty, it will be retried according to the ``retry_pattern``
-    of the job or by :const:`connector.queue.job.RETRY_INTERVAL` if nothing
-    is defined.
-
-    If ``ignore_retry`` is True, the retry counter will not be increased.
-    """
-
-    def __init__(self, msg, seconds=None, ignore_retry=False):
-        super(RetryableJobError, self).__init__(msg)
-        self.seconds = seconds
-        self.ignore_retry = ignore_retry
 
 
 class NetworkRetryableError(RetryableJobError):
     """ A network error caused the failure of the job, it can be retried later.
     """
-
-
-class NothingToDoJob(JobError):
-    """ The Job has nothing to do. """
 
 
 class NoExternalId(RetryableJobError):
@@ -88,7 +61,3 @@ class IDMissingInBackend(JobError):
 
 class ManyIDSInBackend(JobError):
     """Unique key exists many times in backend"""
-
-
-class ChannelNotFound(ConnectorException):
-    """ A channel could not be found """
