+++ connector/security/connector_security.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <data>
 
     <record model="ir.module.category" id="module_category_connector">
         <field name="name">Connector</field>
@@ -13,16 +12,8 @@
         <field name="users" eval="[(4, ref('base.user_root'))]"/>
     </record>
 
-    </data>
-
-    <data noupdate="1">
-
-    <record id="queue_job_comp_rule" model="ir.rule">
-      <field name="name" >Queue job multi-company</field>
-      <field name="model_id" ref="model_queue_job"/>
-      <field name="global" eval="True"/>
-      <field name="domain_force">['|',('company_id','=',False),('company_id','child_of',[user.company_id.id])]</field>
+    <record id="queue_job.group_queue_job_manager" model="res.groups">
+        <field name="implied_ids" eval="[(4, ref('group_connector_manager'))]"/>
     </record>
 
-    </data>
 </odoo>
