+++ connector/security/ir.model.access.csv
@@ -1,5 +1,2 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_connector_queue_job_manager,connector job manager,connector.model_queue_job,connector.group_connector_manager,1,1,1,1
 access_connector_checkpoint_manager,connector checkpoint manager,connector.model_connector_checkpoint,connector.group_connector_manager,1,1,1,1
-access_connector_queue_job_function_manager,connector job functions manager,connector.model_queue_job_function,connector.group_connector_manager,1,1,1,1
-access_connector_queue_job_channel_manager,connector job channel manager,connector.model_queue_job_channel,connector.group_connector_manager,1,1,1,1
