+++ connector/tests/__init__.py
@@ -19,13 +19,10 @@
 #
 ##############################################################################
 
-from . import test_session
-from . import test_event
-from . import test_job
 from . import test_backend
-from . import test_producer
 from . import test_connector
+from . import test_event
 from . import test_mapper
-from . import test_related_action
-from . import test_runner_channels
-from . import test_runner_runner
+from . import test_producer
+from . import test_related_action_binding
+from . import test_session
