--- connector/tests/test_runner_channels.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-import doctest
-from odoo.addons.connector.jobrunner import channels
-
-
-def load_tests(loader, tests, ignore):
-    tests.addTests(doctest.DocTestSuite(channels))
-    return tests
