--- connector/tests/test_runner_runner.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-import doctest
-from odoo.addons.connector.jobrunner import runner
-
-
-def load_tests(loader, tests, ignore):
-    tests.addTests(doctest.DocTestSuite(runner))
-    return tests
